third_party/blink/web_tests/platform/generic/svg/animations/animate-elem-13-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rect1.width.animVal.value is 30
PASS rect1.width.baseVal.value is 30
PASS rect1.width.animVal.value is 165
PASS rect1.width.baseVal.value is 30
PASS rect1.width.animVal.value is 300
PASS rect1.width.baseVal.value is 30
PASS rect1.width.animVal.value is 300
PASS rect1.width.baseVal.value is 30
PASS successfullyParsed is true

TEST COMPLETE
